extract useTypeMetadata into it's own file
